requirements.txt
certifi
aiohttp
websockets
tls-client
websocket-client
deep-translator
colorama
requests
pywin32